Rewrite and update the letsencrypt image
This includes a number of fixes to the way that we set up execute cert
generation for bar-back. We simplify the setup of certbot by using the
apt-get distributed version of the software instead of installing from
github. This also cleans up the scripts for generating the certs and
gets rid of the old logic to recreate pods that use the secret since
we're using a Google load balancer and we don't actually need to restart
any pods to update the certificate.